HackCruz-COS/cos-backend-lite/requirements.txt
# ─── COS Backend Lite — Dependencies ──────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy==2.2.1
scikit-learn==1.6.0
